pyproject.toml
[project]
name = "ui-design"
version = "0.0.1"
authors = []
description = "Open Field Test Analysis"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pandas>=1.5",
    "numpy>=1.22",
    "PyQt6",
    "matplotlib",
    "ruamel.yaml"
]

[tool.setuptools.package-data]
ui_design = ["*.yaml"]
